Bundle Stopify on build
We should have been doing this all along.

With this commit, `yarn run build` produces a bundled version of Stopify:

    var stopify = (function () { ... lots of code ... }();

Now, `./bin/webpack` translates `require('Stopify')` to a reference to the
`stopify variable. Therefore, no more double-Webpack!

Next, we should bundle just the Stopify runtime system and not the compiler.