New --external-rts flag
With this flag, we expect a global variable called "stopify" to refer
to the Stopify runtime system. However, unlike the --webpack flag,
we don't use webpack.

This should help address "double-webpack" issues. It is also necessary to run
as a Cloud Function. Webpack blows up because it tries to open a file for
writing that, as far as I can tell, should not be opened for writing.