Throw exn results from runtime instead of silently continuing.
When delimits are encountered and no delimits are in the queue, delimits are processed immediately. When an exception occurs, the exn is simply returned to the runtime call within the delimit, and is not propogated to the top level, meaning the call to delimit returns normally and the program continues executing. This throws the exn value instead of returning normally.