Fix dependency errors
When another package written in TypeScript imports stopify or
stopify-continuations, we get errors because type annotations are missing for
babel and other modules. The problem here is that a TS package that uses
Stopify requires its type annotations even if Stopify is already built. Stopify
uses some babel types in its interface, which is why we need this.

This is quite hard to test. Using yarn link does not reveal the problem.
The problem only occurs when installing Stopify from NPM.