Use $S.g for globals and remove .externals
The .externals compiler option is gone. Instead, global variables are set as
properties of a global object. You can see an example here:

eeb48eb#diff-8cc0452b1054416799bec2e2b91e6693R45

I don't know if it affects performance (my guess is no). But, it is a
representation change. This approach has two advantages:

a. The global values in stopified code do not have to also be globals in the
   host page. In fact, a global to stopify can be an anonymous value.

b. It enables a simple implementation of eval, which will follow in
   another commit.