AsyncRun.evalAsync method
We already have support for desugaring evals in code. But, this
commit adds support for evaluating new code in the same global
environment as the main program. It requires the global environment
to be an object (which is what the previous commit does). With that,
it is actually really easy to implement.