Stopify 0.7.4
- Some dependabot updates.
- Babel parsing exceptions now handled more gracefully.

In the past, I've simultaneously bumped the versions of all the Stopify
packages. This time, I've only bumped @stopify/stopify to 0.7.4, and
all its dependencies remain at 0.7.3.

However, I am really confused about the semantics of yarn.lock.
We have a shared yarn.lock for all Stopify packages. Dependabot pushed
some updates to this file which are part of this release:

- https://github.com/ocelot-ide/Stopify/commit/070d004f500d6e83cf33b3a25303f409d45079a0
- https://github.com/ocelot-ide/Stopify/commit/824da7065dbe84310afdc9ae130bca0a3256f6be
- https://github.com/ocelot-ide/Stopify/commit/3ac8225e495bed2a783a696addb4852a69970e30
- https://github.com/ocelot-ide/Stopify/commit/9edbf9387c135cd0f326c11391cf783812b3d690
- https://github.com/ocelot-ide/Stopify/commit/f1f8dda826fa937515c6cc83d11cf73531480999

However, since it is a shared yarn.lock, I can't easily tell which of these
dependencies belong to which project. Moreover, they are mostly transitive
dependencies that we have never directly imported! So, do I need to think
more carefully and potentially publish new versions of the other packages?
Or, do I not need to think about this at all?