Propagate runtime config options about stack size
Deep stacks, at least on the lazy impl, seem to be working as long as the
runtime correctly gets the options about stack size.

Eager is still failing with call stack size exceeded with these changes.

The runtimeOpts given to stopifyLocally were not used to configure any starting
stack sizes, this does that configuration on the init() method. I'm not sure
that's the right place, but it's where the estimator and some other config
happen based on the opts field, so it seems reasonable.

Separately, this also adds (optional) params to newRTS to configure stack size
(since those are the parameters the constructors of the various runtimes take).
With just this change (e.g. without the new configuration in init() that sets
fields on rts), you can get deep/lazy working by using e.g.

stopify.newRTS('lazy', 1000, 10)

before calling .run(). But since stopifyLocally takes the runtimeOpts, it seems
right to use them somewhere.